Ensure we use an up to date rubygems version
There were several vulnerabilities disclosed for rubygems recently. This
ensures that we upgrade rubygems to the latest version that should be safe
to use. By always upgrading to the latest one this also means that in the
event of future issues we shouldn't need to make any changes as future
builds will get the latest version as soon as it's available.

Full details on the vulnerability announcement can be found here:
https://www.ruby-lang.org/en/news/2017/08/29/multiple-vulnerabilities-in-rubygems/